pyproject.toml
[project]
name = "bedrock"
version = "0.0.1"
description = "Config-drevet trading-system med asymmetriske setups og cTrader-bot"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [{name = "Snkpipefish"}]

dependencies = [
    # Core
    "pydantic>=2.9,<3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # Data + historikk (SQLite via stdlib — se ADR-002 for hvorfor ikke
    # DuckDB/pyarrow: produksjons-hardwaren mangler SSE4.2/AVX)
    "pandas>=2.2",
    "numpy>=2.2,<2.3",  # SIMD-sensitive, pin upper bound (ADR-002)

    # HTTP / nettverk
    "requests>=2.32",
    "urllib3>=2.0",
    "tenacity>=9.0",

    # Server
    "flask>=3.0",

    # Logging
    "structlog>=24.4",

    # CLI
    "click>=8.1",
]

[project.optional-dependencies]
# Bot-spesifikke (cTrader Open API)
bot = [
    "ctrader-open-api>=0.9.2",
    "twisted>=24.3",
    "protobuf>=6.0",
    "service_identity>=24.0",
]

# Dev-verktøy
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.6",
    "pyright>=1.1.383",
    "yamllint>=1.35",
    "pre-commit>=3.8",
    "commitizen>=3.29",
]

[project.scripts]
bedrock = "bedrock.cli.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bedrock"]

# ─── UV ─────────────────────────────────────────────────────────────────
# `ctrader-open-api==0.9.2` pin-er `protobuf==3.20.1` hardt, mens
# Bedrock krever `protobuf>=6.0`. Lokalt installeres `ctrader-open-api`
# med `--no-deps` (se STATE.md session 41). `uv sync` i CI trenger
# samme override for å løse dep-grafen. Verifisert lokalt at
# ctrader-open-api funker fint mot protobuf 6.x på tross av deres pin.
[tool.uv]
override-dependencies = [
    "protobuf>=6.0",
]

# ─── Ruff ───────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # bugbear
    "C4",   # comprehensions
    "UP",   # pyupgrade
    "SIM",  # simplify
    "TID",  # tidy imports
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # linelength — håndtert av formatter
    # Norske tegn i docstrings/kommentarer/strings (–, ×, ≥, ≤, →, …, å, ø, æ).
    # Ruff flagger disse som "ambiguous" fordi de ligner ASCII-tegn, men de
    # er bevisste valg i prosjektets norske dokumentasjon. False-positives.
    "RUF001",  # ambiguous-unicode-character-string
    "RUF002",  # ambiguous-unicode-character-docstring
    "RUF003",  # ambiguous-unicode-character-comment
    # Stilvalg — ofte mer lesbart originalt
    "SIM102",  # collapsible-if (nested ifs er noen ganger klarere)
    "SIM105",  # suppressible-exception (try/except/pass er tydelig)
    "SIM108",  # if-else-block-instead-of-if-exp (eksplisitt if/else er ok)
    "SIM117",  # multiple-with-statements (nested er noen ganger klarere)
    # Foreslår str-enum / type-alias som ikke gir reell verdi
    "UP040",   # non-pep695-type-alias
    "UP042",   # replace-str-enum
    # Test-mønster: pytest.raises med generisk Exception eller match-pattern
    # som ikke regex-escape-es er ofte hensiktsmessig i tester
    "B017",    # assert-raises-exception
    # Unpack-ignored values: navngitt eksplisitt for klarhet er ofte
    # mer lesbart enn `_` selv om verdien ikke brukes
    "RUF059",  # unused-unpacked-variable
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101",     # assert OK i tester
    "RUF043",   # pytest.raises ambiguous-pattern (test-mønster, ofte ikke regex)
    "RUF034",   # useless-if-else (kan være forklarende i test)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─── Pyright ────────────────────────────────────────────────────────────
[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = "error"
reportMissingTypeStubs = false
venvPath = "."
venv = ".venv"

# ─── Pytest ─────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: tester som tar > 1 sekund",
    "integration: fuller pipeline-tester",
    "backtest: kjører mot historikk",
]

# ─── Coverage ───────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["src/bedrock"]
omit = ["*/tests/*", "*/__main__.py"]

[tool.coverage.report]
precision = 1
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

# ─── Commitizen ─────────────────────────────────────────────────────────
[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
bump_message = "chore(release): bump version to $new_version"
